pyproject.toml
[project]
name = "codon-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "beautifulsoup4==4.13.5",
    "cachecontrol==0.14.3",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cryptography==46.0.1",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "emoji==2.15.0",
    "filetype==1.2.0",
    "firebase-admin==7.1.0",
    "frozenlist==1.7.0",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "google-cloud==0.34.0",
    "google-cloud-core==2.4.3",
    "google-cloud-firestore==2.21.0",
    "google-cloud-storage==3.4.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.75.0",
    "grpcio-status==1.75.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "jiter==0.11.0",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.76",
    "langchain-experimental==0.3.4",
    "langchain-openai==0.3.33",
    "langchain-text-splitters==0.3.11",
    "langdetect==1.0.9",
    "langsmith==0.4.29",
    "lxml==6.0.1",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "marshmallow==3.26.1",
    "mcp[cli]==1.14.1",
    "mdurl==0.1.2",
    "msgpack==1.1.1",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "nltk==3.9.1",
    "numpy==2.3.3",
    "olefile==0.47",
    "openai==1.108.1",
    "orjson==3.11.3",
    "packaging==25.0",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.32.1",
    "psutil==7.1.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pypdf==6.1.0",
    "python-dotenv==1.1.1",
    "python-iso639==2025.2.18",
    "python-magic==0.4.27",
    "python-multipart==0.0.20",
    "python-oxmsg==0.0.2",
    "pyyaml==6.0.2",
    "rapidfuzz==3.14.1",
    "referencing==0.36.2",
    "regex==2025.9.18",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "starlette==0.48.0",
    "tenacity==9.1.2",
    "tiktoken==0.11.0",
    "tqdm==4.67.1",
    "typer==0.19.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "unstructured==0.18.15",
    "unstructured-client==0.42.3",
    "urllib3==2.5.0",
    "uv==0.8.19",
    "uvicorn==0.36.0",
    "webencodings==0.5.1",
    "wrapt==1.17.3",
    "yarl==1.20.1",
    "zstandard==0.25.0",
]
